exercise(6): Abstracted keydown handling into custom hook, implemented via useEffect w/ key dep and ref'd callback
